requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.8.1
asttokens==2.2.1
attrs==23.2.0
azure-core==1.30.1
azure-eventhub==5.12.0
azure-eventhub-checkpointstoreblob-aio==1.1.4
backcall==0.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.3
comm==0.1.2
contourpy==1.2.0
cryptography==42.0.7
cycler==0.12.1
debugpy==1.6.6
decorator==5.1.1
Django==5.0.6
django-rest-framework==0.1.0
djangorestframework==3.15.1
executing==1.2.0
fonttools==4.47.2
frozenlist==1.4.1
gunicorn==22.0.0
idna==3.7
ipykernel==6.21.3
ipython==8.11.0
isodate==0.6.1
jedi==0.18.2
joblib==1.2.0
jupyter_client==8.0.3
jupyter_core==5.2.0
kiwisolver==1.4.5
matplotlib==3.8.2
matplotlib-inline==0.1.6
msrest==0.7.1
multidict==6.0.5
nest-asyncio==1.5.6
nltk==3.8.1
numpy==1.24.2
oauthlib==3.2.2
packaging==23.0
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.2.0
platformdirs==3.1.1
prompt-toolkit==3.0.38
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
Pygments==2.14.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2022.7.1
pyzmq==25.0.1
regex==2022.10.31
requests==2.31.0
requests-oauthlib==2.0.0
scikit-learn==1.4.0
scipy==1.12.0
seaborn==0.13.1
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.2
threadpoolctl==3.2.0
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.11.0
urllib3==2.2.1
wcwidth==0.2.6
yarl==1.9.4
